@@ -13,29 +13,30 @@
     "@stripe/react-stripe-js": "^1.16.5",
     "@stripe/stripe-js": "^1.54.2",
     "@testing-library/jest-dom": "^5.14.1",
-    "@testing-library/react": "^11.2.7",
-    "@testing-library/user-event": "^12.8.3",
+    "@testing-library/react": "^16.0.0",
+    "@testing-library/user-event": "^14.5.2",
     "axios": "^1.6.5",
+    "bootstrap": "^5.3.3",
     "chart.js": "^3.5.1",
     "country-state-city": "^3.0.1",
-    "overlay-navbar": "^1.0.4",
-    "react": "^17.0.2",
+    "overlay-navbar": "^1.2.3",
+    "react": "^18.3.1",
     "react-alert": "^7.0.3",
     "react-alert-template-basic": "^1.0.2",
     "react-chartjs-2": "^3.0.4",
-    "react-dom": "^17.0.2",
+    "react-dom": "^18.3.1",
     "react-helmet": "^6.1.0",
     "react-icons": "^5.0.1",
     "react-js-pagination": "^3.0.2",
     "react-material-ui-carousel": "^2.3.1",
-    "react-redux": "^7.2.4",
-    "react-router-dom": "^5.2.0",
-    "react-scripts": "^5.0.1",
+    "react-redux": "^9.1.2",
+    "react-router-dom": "^6.24.1",
+    "react-scripts": "^3.0.1",
     "react-stripe-js": "^1.1.5",
     "redux": "^4.1.1",
     "redux-devtools-extension": "^2.13.9",
-    "redux-thunk": "^2.3.0",
-    "web-vitals": "^1.1.2",
+    "redux-thunk": "^3.1.0",
+    "web-vitals": "^4.2.1",
     "webfontloader": "^1.6.28"
   },
   "scripts": {
