@@ -1,6 +1,7 @@
 import React, { Fragment } from "react";
+import { Navigate } from "react-router-dom";
 import { useSelector } from "react-redux";
-import { Redirect, Route } from "react-router-dom";
+import { Route } from "react-router-dom";
 
 const ProtectedRoute = ({ isAdmin, component: Component, ...rest }) => {
   const { loading, isAuthenticated, user } = useSelector((state) => state.user);
@@ -12,11 +13,11 @@ const ProtectedRoute = ({ isAdmin, component: Component, ...rest }) => {
           {...rest}
           render={(props) => {
             if (isAuthenticated === false) {
-              return <Redirect to="/login" />;
+              return <Navigate to="/login" />;
             }
 
             if (isAdmin === true && user.role !== "admin") {
-              return <Redirect to="/login" />;
+              return <Navigate to="/login" />;
             }
 
             return <Component {...props} />;
