@@ -7,7 +7,7 @@ import PersonIcon from "@material-ui/icons/Person";
 import ExitToAppIcon from "@material-ui/icons/ExitToApp";
 import ListAltIcon from "@material-ui/icons/ListAlt";
 import ShoppingCartIcon from "@material-ui/icons/ShoppingCart";
-import { useHistory } from "react-router-dom";
+import { useNavigate } from "react-router-dom";
 import { useAlert } from "react-alert";
 import { logout } from "../../../actions/userAction";
 import { useDispatch, useSelector } from "react-redux";
@@ -16,7 +16,7 @@ const UserOptions = ({ user }) => {
   const { cartItems } = useSelector((state) => state.cart);
 
   const [open, setOpen] = useState(false);
-  const history = useHistory();
+  const navigate = useNavigate();
   const alert = useAlert();
   const dispatch = useDispatch();
 
@@ -44,17 +44,17 @@ const UserOptions = ({ user }) => {
   }
 
   function dashboard() {
-    history.push("/admin/dashboard");
+    navigate("/admin/dashboard");
   }
 
   function orders() {
-    history.push("/orders");
+    navigate("/orders");
   }
   function account() {
-    history.push("/account");
+    navigate("/account");
   }
   function cart() {
-    history.push("/cart");
+    navigate("/cart");
   }
   function logoutUser() {
     dispatch(logout());
