@@ -1,5 +1,5 @@
 import React from "react";
-import ReactDOM from "react-dom";
+import  ReactDOM from "react-dom/client";
 import App from "./App";
 import { Provider } from "react-redux";
 import store from "./store";
@@ -13,11 +13,11 @@ const options = {
   transition: transitions.SCALE,
 };
 
-ReactDOM.render(
+const root = ReactDOM.createRoot(document.getElementById('root'))
+root.render(
   <Provider store={store}>
     <AlertProvider template={AlertTemplate} {...options}>
       <App />
     </AlertProvider>
-  </Provider>,
-  document.getElementById("root")
+  </Provider>
 );
