@@ -1,76 +1,65 @@
-import { createStore, combineReducers, applyMiddleware } from "redux";
-import thunk from "redux-thunk";
-import { composeWithDevTools } from "redux-devtools-extension";
+import { createStore, combineReducers, applyMiddleware } from "redux"
+import {thunk} from "redux-thunk"
+import { composeWithDevTools } from "redux-devtools-extension"
 import {
-  newProductReducer,
-  newReviewReducer,
-  productDetailsReducer,
-  relatedProductsReducer,
-  productReducer,
-  productReviewsReducer,
-  productsReducer,
-  reviewReducer,
-
-} from "./reducers/productReducer";
-
+    newProductReducer,
+    newReviewReducer,
+    productDetailsReducer,
+    productReducer,
+    productReviewsReducer,
+    productsReducer,
+    reviewReducer,
+} from "./reducers/productReducer"
 import {
-  allUsersReducer,
-  forgotPasswordReducer,
-  profileReducer,
-  userDetailsReducer,
-  userReducer,
-} from "./reducers/userReducer";
-
-import { cartReducer } from "./reducers/cartReducer";
+    allUsersReducer,
+    forgotPasswordReducer,
+    profileReducer,
+    userDetailsReducer,
+    userReducer,
+} from "./reducers/userReducer"
+import { cartReducer } from "./reducers/cartReducer"
 import {
-  allOrdersReducer,
-  myOrdersReducer,
-  newOrderReducer,
-  orderDetailsReducer,
-  orderReducer,
-  orderCancelReducer
-} from "./reducers/orderReducer";
+    allOrdersReducer,
+    myOrdersReducer,
+    newOrderReducer,
+    orderDetailsReducer,
+    orderReducer,
+} from "./reducers/orderReducer"
 
 const reducer = combineReducers({
-  products: productsReducer,
-  productDetails: productDetailsReducer,
-  relatedProducts:relatedProductsReducer,
-  user: userReducer,
-  profile: profileReducer,
-  forgotPassword: forgotPasswordReducer,
-  cart: cartReducer,
-  newOrder: newOrderReducer,
-  myOrders: myOrdersReducer,
-  orderDetails: orderDetailsReducer,
-  newReview: newReviewReducer,
-  newProduct: newProductReducer,
-  product: productReducer,
-  allOrders: allOrdersReducer,
-  order: orderReducer,
-  cancelOrder:orderCancelReducer,
-  allUsers: allUsersReducer,
-  userDetails: userDetailsReducer,
-  productReviews: productReviewsReducer,
-  review: reviewReducer,
-});
+    products: productsReducer,//all products with filter,search and pagination
+    productDetail: productDetailsReducer,
+    user: userReducer,
+    profile: profileReducer,
+    forgotPassword: forgotPasswordReducer,
+    cart: cartReducer,
+    newOrder: newOrderReducer,
+    myOrders: myOrdersReducer,
+    orderDetails: orderDetailsReducer,
+    newReview: newReviewReducer,
+    newProduct: newProductReducer,
+    product: productReducer,
+    allOrders: allOrdersReducer,
+    order: orderReducer,
+    allUsers: allUsersReducer,
+    userDetails: userDetailsReducer,
+    productReviews: productReviewsReducer,
+    review: reviewReducer,
+})
 
 let initialState = {
-  cart: {
-    cartItems: localStorage.getItem("cartItems")
-      ? JSON.parse(localStorage.getItem("cartItems"))
-      : [],
-    shippingInfo: localStorage.getItem("shippingInfo")
-      ? JSON.parse(localStorage.getItem("shippingInfo"))
-      : {},
-  },
-};
+    cart: {
+        cartItems: localStorage.getItem("cartItems")
+            ? JSON.parse(localStorage.getItem("cartItems"))
+            : [],
+        shippingInfo: localStorage.getItem("shippingInfo")
+            ? JSON.parse(localStorage.getItem("shippingInfo"))
+            : {},
+    },
+}
 
-const middleware = [thunk];
+const middleware = [thunk]
 
-const store = createStore(
-  reducer,
-  initialState,
-  composeWithDevTools(applyMiddleware(...middleware))
-);
+const store = createStore(reducer, initialState, composeWithDevTools(applyMiddleware(...middleware)))
 
-export default store;
+export default store
